+++ lib/src/dart_ping_base.dart
@@ -6,7 +6,10 @@ import 'package:dart_ping/src/native_ping.dart';
 abstract class Ping {
   /// Creates an appropriate Ping instance for the detected platform
   factory Ping(String host,
-          {int count, double interval, double timeout, bool ipv6}) =>
+          {int? count,
+          double interval = 1.0,
+          double timeout = 2.0,
+          bool ipv6 = false}) =>
       getPing(host, count, interval, timeout, ipv6);
 
   /// Stream of [PingData] which spawns a ping process on listen and