+++ lib/src/native_ping.dart
@@ -6,7 +6,7 @@ import 'package:dart_ping/src/ping/mac_ping.dart';
 import 'package:dart_ping/src/ping/windows_ping.dart';
 
 Ping getPing(
-    String host, int count, double interval, double timeout, bool ipv6) {
+    String host, int? count, double interval, double timeout, bool ipv6) {
   switch (Platform.operatingSystem) {
     case 'android':
     case 'fuchsia':